+++ auctionapp/templates/item/category.html
@@ -15,7 +15,7 @@
         {% for item in item_list %}
         <div>
             <h4> {{ item.name }} </h4>
-            <a href="{{ url_for('product', item_id=item.id) }}" class="item-link"></a>
+            <a href="{{ url_for('item.product', item_id=item.id) }}" class="item-link"></a>
             {% if item.photo %}
                     <img src="{{ url_for('static', filename=item.photo) }}" alt="">
             {% else %}
