--- auctionapp/templates/site/category.html
@@ -1,45 +1,0 @@
-{% extends "base.html" %}
-{% block content %}
-
-<main>
-    <h2>{{ page_title }} </h2>
-    <div class="main-article">
-        <p>
-            Lorem ipsum dolor sit amet, consectetur adipisicing elit. Accusantium atque delectus,
-            dolorem ea facere fugiat inventore itaque modi molestias nihil nobis nostrum officia
-            quibusdam quis repellendus sapiente, ut vitae voluptatum.
-        </p>
-    </div>
-    <article>
-    <div class="items">
-        {% for item in item_list %}
-        <div>
-            <h4> {{ item.name }} </h4>
-            <a href="{{ url_for('product', item_id=item.id) }}" class="item-link"></a>
-            {% if item.photo %}
-                    <img src="{{ url_for('static', filename=item.photo) }}" alt="">
-            {% else %}
-                    <img src="{{ url_for('static', filename='/images/items/default.jpg') }}" alt="">    
-            {% endif %}
-            <p> {{ item.description }}</p>
-            <p> Год: <span>{{ item.year }}</span></p>
-            {% if item.status == 'success_sale' %}
-                    <h3>Продан</h3>
-                    <p>Последняя цена: <span>{{ item.last_price }}</span></p>
-            {% else %}
-                    {% if item.status == 'failed_sale' %}
-                            <h4>Аукцион не состоялся</h4>
-                            <p>Последняя цена: <span>{{ item.nom_price }} </span></p>
-                    {% else %}
-                        <h4>Стартовая цена:</h4>
-                        <h5>{{ item.nom_price }}
-                    {% endif %}
-            {% endif %}
-        </div>
-        {% endfor %}
-
-    </div>
-</article>
-</main>
-
-{% endblock %}
