+++ auctionapp/templates/site/popular.html
@@ -14,7 +14,7 @@
     <div class="items">
         {% for item in items_list %}
         <div>
-            <a href="{{ url_for('product', item_id=item.id) }}" class="item-link"></a>
+            <a href="{{ url_for('item.product', item_id=item.id) }}" class="item-link"></a>
             {% if item.photo %}
                 <img src="{{ url_for('static', filename=item.photo) }}" alt="">
             {% else %}
